rag-chatbot/requirements.txt
chromadb==1.0.10
streamlit==1.41.1
onnxruntime==1.22.0
PyMuPDF==1.26.0
huggingface-hub==0.32.1

pydantic==2.11.5
pydantic_core==2.33.2

python-dotenv==1.1.0
requests==2.32.5
tokenizers==0.21.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
zipp==3.21.0
rich==13.7.1
tenacity==8.2.3
protobuf>=5,<6

# LangChain / LangGraph
langgraph==1.0.3
langchain==1.0.7
langchain-community==0.4.1

# Langfuse + OpenAI
langfuse==3.10.0
openai==2.7.2
backoff==2.2.1
httpx==0.27.2

opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0


grandalf==0.8
